refactor: adding `setupCors` function
